requirements.txt
flask==0.12.4
gunicorn==19.6.0
pyyaml==3.12
python-slugify==1.2.1
Flask-PyMongo==0.5.1
ujson==1.35
flask-compress==1.4.0
click==6.7
Babel==2.6.0
Flask-Caching==1.4.0
pymongo==3.7.0
azure-keyvault==1.1.0
msrestazure==0.5.1
